pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag_app"
version = "0.1.0"
description = "RAG assistant service (placeholder)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "httpx",
  "pydantic-settings",
  "uvicorn",
  "pydantic",
  "typer",
  "python-dotenv",
  "python-docx",
  "pillow",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "mypy",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.ruff.lint.isort]
known-first-party = ["rag_app"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true

[tool.pytest.ini_options]
pythonpath = ["."]
